Added 6 callback events:  onPreSave, onPostSave, onSaveError, onPreRestore, onPostRestore, onRestoreError
All 6 events are public properties of the TinyAutoSave plugin, and they are assigned String values in exactly the same way that TinyMCE events are configured.  The callback function can either be a global function (onPostSave: "myFunction") or a dot-notation function with scope (onPostSave: "myObject.myMethod").

The callbacks use the internal TinyMCE callback methods, so they should be highly compatible with TinyMCE programming.  The editor instance is the context of the callback (set as 'this').

Documentation is included for the events.

Also added error trapping during both save and restore.  Some users were experiencing errors due to low disk space, and the plugin would throw an error.  There are now blanket try/catch wrappers around both save and restore, so any errors encountered should be successfully trapped.  The onSaveError and onRestoreError events can be used to notify the user, if desired.